11-10-2026/requirements.txt
boto3>=1.34.0
pyyaml>=6.0
flask>=3.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
psutil>=5.9.0
